+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Generic filter: options to match if a field is/isn't defined.
 - Excellon drill: added `route_mode_for_oval_holes` option.
 - Default global `dir` option.
+- Pattern to expand the variant name: %V
+- PCB PDF Print: mechanism to change the block title. (#102)
 
 ### Changed
 - Internal BoM: now components with different Tolerance, Voltage, Current